requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.4
openai>=1.68.2
python-dotenv==1.0.0
sqlite-utils==3.35.2
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.4.15